refactor:http2 fallback handler
