+++ zio-http/src/main/scala/zhttp/service/server/ServerChannelInitializer.scala
@@ -1,21 +1,8 @@
 package zhttp.service.server
 
-import io.netty.channel.{
-  ChannelHandlerContext => JChannelHandlerContext,
-  SimpleChannelInboundHandler => JSimpleChannelInboundHandler,
-}
 import io.netty.handler.codec.http.HttpServerUpgradeHandler.{UpgradeCodecFactory => JUpgradeCodecFactory}
-import io.netty.handler.codec.http.{
-  HttpMessage => JHttpMessage,
-  HttpServerCodec => JHttpServerCodec,
-  HttpServerKeepAliveHandler => JHttpServerKeepAliveHandler,
-  HttpServerUpgradeHandler => JHttpServerUpgradeHandler,
-}
-import io.netty.handler.codec.http2.{
-  Http2CodecUtil => JHttp2CodecUtil,
-  Http2FrameCodecBuilder => JHttp2FrameCodecBuilder,
-  Http2ServerUpgradeCodec => JHttp2ServerUpgradeCodec,
-}
+import io.netty.handler.codec.http.{HttpServerCodec => JHttpServerCodec, HttpServerKeepAliveHandler => JHttpServerKeepAliveHandler, HttpServerUpgradeHandler => JHttpServerUpgradeHandler}
+import io.netty.handler.codec.http2.{Http2CodecUtil => JHttp2CodecUtil, Http2FrameCodecBuilder => JHttp2FrameCodecBuilder, Http2ServerUpgradeCodec => JHttp2ServerUpgradeCodec}
 import io.netty.util.{AsciiString => JAsciiString}
 import zhttp.core._
 import zhttp.service.Server.Settings
@@ -51,29 +38,12 @@ object ServerChannelInitializer {
     c: JChannel,
     settings: Settings[R, Throwable],
   ) = if (settings.enableHttp2) {
-
-    //TODO: add an encryptedmessagefilter
     val p           = c.pipeline
     val sourceCodec = new JHttpServerCodec
-    p.addLast(EncryptedMessageFilter(httpH, settings))
+    p.addLast(ENCRYPTION_FILTER_HANDLER,EncryptedMessageFilter(httpH, settings))
     p.addLast(SERVER_CODEC_HANDLER, sourceCodec)
     p.addLast(CLEAR_TEXT_HTTP2_HANDLER, new JHttpServerUpgradeHandler(sourceCodec, upgradeCodecFactory(http2H)))
-    p.addLast(
-      CLEAR_TEXT_HTTP2_FALLBACK_HANDLER,
-      new JSimpleChannelInboundHandler[JHttpMessage]() {
-        @throws[Exception]
-        override protected def channelRead0(ctx: JChannelHandlerContext, msg: JHttpMessage): Unit = { // If this handler is hit then no upgrade has been attempted and the client is just talking HTTP.
-          val pipeline = ctx.pipeline
-          val thisCtx  = pipeline.context(this)
-          pipeline
-            .addAfter(thisCtx.name(), OBJECT_AGGREGATOR, new JHttpObjectAggregator(settings.maxRequestSize))
-            .addAfter(OBJECT_AGGREGATOR, HTTP_REQUEST_HANDLER, httpH)
-            .replace(this, HTTP_KEEPALIVE_HANDLER, new JHttpServerKeepAliveHandler)
-          ctx.fireChannelRead(msg)
-          ()
-        }
-      },
-    )
+    p.addLast(CLEAR_TEXT_HTTP2_FALLBACK_HANDLER,ClearTextHttp2FallbackHandler(httpH,settings))
     ()
   } else {
     c.pipeline()
