+++ zio-http/src/main/scala/zhttp/service/server/ServerChannelInitializerUtil.scala
@@ -1,0 +1,7 @@
+package zhttp.service.server
+
+final class ServerChannelInitializerUtil {
+
+
+
+}
